application commit 1) act filter tayyor buldi

+++ Domain/src/main/java/uz/idea/domain/models/userInfo/Data.kt
@@ -8,5 +8,5 @@ data class Data(
     val full_name: String,
     val phone: String,
     val tariff: Tariff,
-    val tin: Int
+    val tin: Long
 )
